+++ watchtower-app/src/panels/iframe-panel.js
@@ -29,6 +29,7 @@ export class IframePanel extends LitElement {
     }
 
     refresh() {
+        if (this.config.noRefresh === true) return
         this.renderRoot.querySelector('iframe').src += ''
     }
 }
@@ -36,6 +37,8 @@ export class IframePanel extends LitElement {
 export class IframePanelConfig extends LitElement {
     static properties = {
         url: { type: String, state: true },
+        noBorder: { type: Boolean, state: true },
+        noRefresh: { type: Boolean, state: true },
     }
 
     constructor() {
@@ -45,8 +48,18 @@ export class IframePanelConfig extends LitElement {
 
     render() {
         return html`
-            <label for="url">URL to load:</label>
-            <input type="url" id="url" required="true" type="url" pattern="https?://.+" placeholder="Enter URL to load" autocomplete="off" @change=${event => this.url = event.target.value}/>
+            <section>
+                <label for="url">URL to load:</label>
+                <input type="url" id="url" required="true" type="url" pattern="https?://.+" placeholder="Enter URL to load" autocomplete="off" @change=${event => this.url = event.target.value}/>
+            </section>
+            <section>
+                <label><input value="true" type="checkbox"
+                    @change=${this.onNoBorderChecked}
+                > Hide tile border</label>
+                <label><input value="true" type="checkbox"
+                    @change=${this.onNoRefreshChecked}
+                > Disable auto-refresh</label>
+            </section>
         `
     }
 
@@ -58,7 +71,20 @@ export class IframePanelConfig extends LitElement {
         setTimeout(() => this.renderRoot.querySelector('#url').focus(), 0)
     }
 
+    onNoBorderChecked(event) {
+        this.noBorder = event.target.checked
+    }
+    
+    onNoRefreshChecked(event) {
+        this.noRefresh = event.target.checked
+    }
+
     decorateConfig(config) {
-        return { ...config, url: this.url }
+        return {
+            ...config,
+            url: this.url,
+            noBorder: this.noBorder,
+            noRefresh: this.noRefresh
+        }
     }
 }